Are you fucking kidding me
I had completely missed a closing brace here and it somehow didn't error when it was supposed to. I swear to god that was the reason it didn't work. I didn't really inspect the code and insteadfgot claude.ai to do it for me cuz I didn't really feel like doing all that nonsense. I am so relieved by that. However that doesn't mean it fixes it though cuz the closing brace was already there in the git version.